+++ NPCMapLocations/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿﻿﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* The minimap is no longer shifted to avoid being drawn over viewport black bars (e.g. at the bus stop).
+
 ## 2.11.3
 Released 30 March 2023 for SMAPI 3.14.0 or later. Updated by Pathoschild.